add tests for metric helpers getting class label values
